Disable only identifier widget instead of whole inspector
